Better guide and cleanup
