+++ README.md
@@ -1,8 +1,8 @@
-## Trajectory planning visualization tool
+# Trajectory planning visualization tool
 
-### Description
+## Description
 
-This tool is a companion to the `dabry` module (<https://github.com/bschnitzler/dabry>)
+This tool is a companion to the [Dabry module](<https://github.com/bschnitzler/dabry>).
 
 It handles the visualisation of trajectory optimization problems 
 and the different objects at stake:
@@ -14,22 +14,69 @@ and the different objects at stake:
 It can handle time-varying winds, 2D planar as well as spherical (Earth)
 problems.
 
-### Getting started
+## Installation
 
-1) Write the path to the `dabry` module in the `DABRYPATH` file.
-2) Run the script `run.sh` using for instance
-```sh
-bash run.sh
+Follow the instructions for the Dabry module.
+You should now have a working directory with structure:
+```
+wdir
+└── dabry
+└── env
 ```
 
-### Examples
+`cd` into this working directory and clone the dabryvisu repo 
+```sh
+git clone [repo-url]
+```
+
+Then activate your environment 
+```sh
+source env/bin/activate
+```
+
+Install `dabryvisu` 
+```sh
+pip install ./dabry-visu
+``` 
+
+The variable `DABRYPATH` shall point to the cloned `dabry` repository.
+This is the case if you followed Dabry's installation instructions.
+
+## Running the tool
+
+Common way to run the tool is
+```sh
+python -m dabryvisu
+```
+
+To run the visualization of the latest case, use
+
+```sh
+python -m dabryvisu -l
+```
+
+To create an animation from a case, use
+
+```sh
+python -m dabryvisu -m
+```
+
+To see all options, print help
+
+```sh
+python -m dabryvisu -h
+```
+
+## Examples
 
 Here are some previews of the tool
 
 - Resolution with extremal field
 
-![](https://github.com/bschnitzler/dabry-visu/blob/main/res/eft.gif)
+![](res/eft.gif)
 
-- Resolution with front tracking (interface to ToolboxLS)
+- Resolution with front tracking (interface to [ToolboxLS](https://www.cs.ubc.ca/~mitchell/ToolboxLS))
 
-![](https://github.com/bschnitzler/dabry-visu/blob/main/res/rft.gif)
+![](res/rft.gif)
+
+More examples can be found on the dedicated [website](https://dabry-navigation.github.io).
