--- requirements.txt
@@ -1,14 +1,0 @@
-basemap==1.3.6
-easygui==0.98.3
-eel==0.15.1
-geopy==2.3.0
-h5py
-ipython==8.9.0
-ipywidgets==7.6.5
-matplotlib==3.4.3
-mpld3==0.5.9
-numpy
-pyproj==3.4.1
-PyQt5
-scipy
-tqdm==4.62.3
