--- src/dabryvisu/font_config.py
@@ -1,24 +1,0 @@
-class FontsizeConf:
-
-    def __init__(self):
-        self.fontsize = 18
-        self.axes_titlesize = 24
-        self.axes_labelsize = 22
-        self.xtick_labelsize = 22
-        self.ytick_labelsize = 22
-        self.legend_fontsize = 15
-        self.button_fontsize = 14
-        self.timedisp_major = 20
-        self.timedisp_minor = 14
-        self.font_family = 'lato'
-        self.mathtext_fontset = 'cm'
-
-    # def __init__(self):
-    #     self.fontsize = 10
-    #     self.axes_titlesize = 10
-    #     self.axes_labelsize = 8
-    #     self.xtick_labelsize = 7
-    #     self.ytick_labelsize = 7
-    #     self.legend_fontsize = 10
-    #     self.font_family = 'lato'
-    #     self.mathtext_fontset = 'cm'
